# Fuse input copy in cuda graph replay