--- a/sgl-kernel/csrc/common_extension.cc
+++ b/sgl-kernel/csrc/common_extension.cc
@@ -227,6 +227,14 @@ TORCH_LIBRARY_FRAGMENT(sgl_kernel, m) {
   m.def("segment_packbits(Tensor x, Tensor input_indptr, Tensor output_indptr, Tensor! y, int cuda_stream) -> ()");
   m.impl("segment_packbits", torch::kCUDA, &segment_packbits);
 
+  m.def(
+      "copy_cuda_graph_replay_inputs(Tensor! input_ids_dst, Tensor! seq_lens_dst, Tensor! extend_seq_lens_dst, Tensor! "
+      "out_cache_loc_dst, Tensor! positions_dst, Tensor! req_pool_indices_dst, Tensor! accept_length_dst, Tensor! "
+      "hidden_states_dst, Tensor input_ids_src, Tensor seq_lens_src, Tensor extend_seq_lens_src, Tensor "
+      "out_cache_loc_src, Tensor positions_src, Tensor req_pool_indices_src, Tensor accept_length_src, Tensor "
+      "hidden_states_src, int num_tokens, int raw_bs, int hidden_size) -> ()");
+  m.impl("copy_cuda_graph_replay_inputs", torch::kCUDA, &copy_cuda_graph_replay_inputs);
+
   /*
    * From FlashInfer
    */