--- a/sgl-kernel/python/sgl_kernel/__init__.py
+++ b/sgl-kernel/python/sgl_kernel/__init__.py
@@ -68,6 +68,7 @@ from sgl_kernel.sampling import (
 )
 from sgl_kernel.speculative import (
     build_tree_kernel_efficient,
+    copy_cuda_graph_replay_inputs,
     segment_packbits,
     tree_speculative_sampling_target_only,
     verify_tree_greedy,