<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Add fused copy kernel for 3 cuda graph replay stage to reduce kernel launch overhead.

## Modifications
1. src and dst tensor dtype alignment (except `input_ids`, it's int32 for draft extend, draft decode, and int64 for target verify. It's hard to track where is the initial setting of the dtype)
2. fuse fill, zero and copy for serval inputs into 1 cuda kernel. Here we use cuda kernel since it can have less launch overhead than triton kernel out of cuda graph.
3. compatible with 3 cuda graph's inputs
4. dispatch for different dtypes

## Profile
main branch:
<img width="473" alt="image" src="https://github.com/user-attachments/assets/c1a0a00d-0845-4cd9-93a5-99aa1a8e60ae" />
<img width="423" alt="image" src="https://github.com/user-attachments/assets/12b0da62-6d35-4ebc-b05f-d6c08217b2a4" />


This PR:
<img width="306" alt="image" src="https://github.com/user-attachments/assets/da8768cc-16d2-4d49-9f03-fe9fd5405ecc" />
<img width="240" alt="image" src="https://github.com/user-attachments/assets/da161713-fc39-414c-b80e-1d54302f10e7" />


## Benchmark
```
python3 -m sglang.launch_server --model-path meta-llama/Meta-Llama-3-8B-Instruct --speculative-algorithm EAGLE --speculative-draft-model-path lmsys/sglang-EAGLE-LLaMA3-Instruct-8B --speculative-num-steps 2 --speculative-eagle-topk 2 --speculative-num-draft-tokens 3 --trust-remote-code --dtype float16 --attention-backend fa3

curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 1000 --random-output-len 1000 --random-range-ratio 1 --num-prompts 5 --max-concurrency 1 --output-file result.jsonl
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 1000 --random-output-len 1000 --random-range-ratio 1 --num-prompts 20 --max-concurrency 4 --output-file result.jsonl
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 1000 --random-output-len 1000 --random-range-ratio 1 --num-prompts 80 --max-concurrency 16 --output-file result.jsonl
curl http://127.0.0.1:30000/flush_cache
python3 -m sglang.bench_serving --backend sglang-oai  --dataset-name random --random-input-len 1000 --random-output-len 1000 --random-range-ratio 1 --num-prompts 160 --max-concurrency 32 --output-file result.jsonl
```

main branch:
```
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            242.976 |             242.976 |         30.030 |           32.125 |        34.665 |          4.089 |            3.962 |         4.807 |               242.976 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            881.925 |             881.925 |         46.290 |           36.787 |       104.169 |          4.281 |            4.290 |         5.142 |               220.481 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |           2688.907 |            2688.907 |        102.308 |           37.939 |       453.306 |          5.115 |            5.066 |         6.602 |               168.057 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |            32.000 |           4747.085 |            4747.085 |        175.715 |           38.818 |       981.174 |          6.117 |            6.005 |         8.497 |               148.346 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```
This PR:
```
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            245.437 |             245.437 |         45.779 |           32.865 |       104.632 |          4.031 |            3.906 |         4.772 |               245.437 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            906.386 |             906.386 |         57.605 |           36.456 |       183.936 |          4.148 |            4.223 |         5.071 |               226.596 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |           2727.523 |            2727.523 |         90.373 |           37.744 |       388.891 |          5.046 |            5.000 |         6.537 |               170.470 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |            32.000 |           4758.935 |            4758.935 |        136.109 |           38.731 |       777.836 |          6.142 |            6.005 |         8.616 |               148.717 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.