@ispobock 2025-06-22
> in UT, this PR raise "ImportError: cannot import name 'copy_cuda_graph_replay_inputs' from 'sgl_kernel' (/usr/local/lib/python3.10/dist-packages/sgl_kernel/init.py)", please fix this

Thanks for reviewing! Yes, for normal kernel development workflow in SGLang. We should update the kernel code first, then apply to the python code afther the new kernel version released. This PR just shows some results and code changes, it's not going to merge. 
Since the EAGLE part will be refactored recently, we may don't need this fusion.